pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartcommit"
version = "0.1.0"
description = "AI-powered git commit message generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "gitpython>=3.1.40",
    "anthropic>=0.18.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
]

[project.scripts]
smartcommit = "smartcommit.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]